requirements.txt
pyaudio
pydub
openai-whisper
noisereduce